chore(meta): move BooleanExpression utils to separate file
